backend/requirements/base.txt
Django==2.0.2
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-extensions==1.9.9
django-rest-swagger==2.1.2
djangorestframework==3.7.7
Pillow==5.0.0
psycopg2==2.7.3.2

